+++ tests/Util.hs
@@ -84,6 +84,14 @@ expectEq t file line context x y =
   [show x <> " equals "     <> show y]
   [show x <> " is not equal to " <> show y]
 
+expectNe :: (Eq a, Show a, Show b) =>
+            TestEnv -> String -> Integer -> b -> a -> a -> IO ()
+expectNe t file line context x y =
+  check t (x /= y)
+  [showContext file line context]
+  [show x <> " is not equal to " <> show y]
+  [show x <> " equals "     <> show y]
+
 expectNear :: (Num a, Ord a, Show a, Show b) =>
               TestEnv -> String -> Integer -> b -> a -> a -> a -> IO ()
 expectNear t file line context x y diff =