+++ changelog.md
@@ -13,6 +13,8 @@ Changelog for the [`directory`][1] package
   * Changed `canonicalizePath` to dereference symbolic links even if it points
     to a file and is not the last path segment
 
+  * On Windows, `canonicalizePath` now canonicalizes the letter case too
+
 ## 1.2.7.1 (November 2016)
 
   * Don't abort `removePathForcibly` if files or directories go missing.